Lots of miscellaneous tidy (#377)
* Tidy some DataFunc things
- Adjusted DefDataFunc macro and applied it to all registered functions
- Normalized parameter name to `array`
- Remove explicit DataNode constructions where not necessary

* Tidy a couple objects.json loading things

* A couple DataFunc inlines

* Better names for UIFontImporter font size inlines

* Some DataArray.cpp cleanup

* Fix a few clangd warnings

* Fix MILO_TRY/CATCH comments

* Set default array parameter for DataNode value retrieval functions

* Tidy up handler methods, and fix some handle cases

* Add new handle macros for further tidying, and tidy HANDLE_ACTION family
HANDLE_ACTION now requires the comma operator for multiple actions

* Fix a couple retail DataFunc things
- PrintToStream doesn't actually exist unfortunately
- DataLowestBit uses an inline

* Fix StlNodeAlloc conversion errors in retail

+++ src/system/bandobj/FingerShape.cpp
@@ -30,7 +30,7 @@ FingerShape::FingerShape(RndDir* dir) : mLastState(new RGState()), mFretNumberSh
     mFretNumberShowAnim = dir->Find<RndAnimatable>("fret_number_show.anim", true);
     mFretNumberPositionAnim = dir->Find<RndAnimatable>("fret_number_pos.tnm", true);
     mFretNumberText = dir->Find<RndText>("fret_numbers_chord.txt", true);
-    mAnimPeriod = dir->Property("anim_period", true)->Float(0);
+    mAnimPeriod = dir->Property("anim_period", true)->Float();
 }
 
 FingerShape::~FingerShape(){
