+++ src/system/ui/UIListDir.cpp
@@ -191,7 +191,7 @@ void UIListDir::DrawShowing(){
         DrawWidgets(mTestState, unk1fc, WorldXfm(), mTestComponentState, 0, false);
     }
     else
-#endif 
+#endif
         RndDir::DrawShowing();
 }
 
@@ -215,17 +215,17 @@ bool UIListDir::IsActive(int i) const {
 }
 
 BEGIN_PROPSYNCS(UIListDir)
-    SYNC_PROP_SET(orientation, mOrientation, mOrientation = (UIListOrientation)_val.Int(0))
+    SYNC_PROP_SET(orientation, mOrientation, mOrientation = (UIListOrientation)_val.Int())
     SYNC_PROP(fade_offset, mFadeOffset)
     SYNC_PROP(element_spacing, mElementSpacing)
     SYNC_PROP(scroll_highlight_change, mScrollHighlightChange)
     SYNC_PROP(test_mode, mTestMode)
     SYNC_PROP(test_num_data, mTestNumData)
     SYNC_PROP(test_gap_size, mTestGapSize)
-    SYNC_PROP_SET(test_num_display, mTestState.mNumDisplay, mTestState.SetNumDisplay(_val.Int(0), true))
-    SYNC_PROP_SET(test_grid_span, mTestState.mGridSpan, mTestState.SetGridSpan(_val.Int(0), true))
-    SYNC_PROP_SET(test_scroll_time, mTestState.Speed(), mTestState.SetSpeed(_val.Float(0)))
-    SYNC_PROP_SET(test_list_state, mTestComponentState, mTestComponentState = (UIComponent::State)_val.Int(0))
+    SYNC_PROP_SET(test_num_display, mTestState.mNumDisplay, mTestState.SetNumDisplay(_val.Int(), true))
+    SYNC_PROP_SET(test_grid_span, mTestState.mGridSpan, mTestState.SetGridSpan(_val.Int(), true))
+    SYNC_PROP_SET(test_scroll_time, mTestState.Speed(), mTestState.SetSpeed(_val.Float()))
+    SYNC_PROP_SET(test_list_state, mTestComponentState, mTestComponentState = (UIComponent::State)_val.Int())
     SYNC_PROP_MODIFY(test_disable_elements, mTestDisableElements, Reset())
     SYNC_SUPERCLASS(RndDir)
 END_PROPSYNCS
